requirements.txt
# Core async HTTP
aiohttp==3.9.5

# Google Cloud
google-auth==2.29.0
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.1
google-cloud-dlp==3.15.0
google-cloud-secret-manager==2.20.0

# Google Cloud Healthcare FHIR API
google-cloud-healthcare==1.11.0

# Vertex AI / Gemini
google-cloud-aiplatform==1.52.0
vertexai==1.52.0

# Data validation
pydantic==2.7.1

# Environment variable loading for local dev
python-dotenv==1.0.1

# Utilities
uuid==1.30